Use `shouldRemove...` functions, needed to downgrade type union
