+++ rules/downgrade/src/Contract/Rector/DowngradeParamDeclarationRectorInterface.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\Downgrade\Contract\Rector;
+
+use PhpParser\Node\Param;
+
+interface DowngradeParamDeclarationRectorInterface
+{
+    /**
+     * Indicate if the parameter must be removed
+     */
+    public function shouldRemoveParamDeclaration(Param $param): bool;
+}
