+++ rules/downgrade/src/Contract/Rector/DowngradeReturnDeclarationRectorInterface.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\Downgrade\Contract\Rector;
+
+use PhpParser\Node\FunctionLike;
+
+interface DowngradeReturnDeclarationRectorInterface
+{
+    /**
+     * Indicate if the return declaration must be removed
+     */
+    public function shouldRemoveReturnDeclaration(FunctionLike $functionLike): bool;
+}
