+++ rules/downgrade/src/Rector/FunctionLike/AbstractDowngradeReturnDeclarationRector.php
@@ -1,0 +1,68 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\Downgrade\Rector\FunctionLike;
+
+use PhpParser\Node;
+use PhpParser\Node\FunctionLike;
+use PhpParser\Node\Stmt\ClassMethod;
+use PhpParser\Node\Stmt\Function_;
+use Rector\BetterPhpDocParser\PhpDocInfo\PhpDocInfo;
+use Rector\Downgrade\Contract\Rector\DowngradeReturnDeclarationRectorInterface;
+use Rector\Downgrade\Rector\Property\AbstractDowngradeRector;
+use Rector\NodeTypeResolver\Node\AttributeKey;
+
+abstract class AbstractDowngradeReturnDeclarationRector extends AbstractDowngradeRector implements DowngradeReturnDeclarationRectorInterface
+{
+    /**
+     * @return string[]
+     */
+    public function getNodeTypes(): array
+    {
+        return [Function_::class, ClassMethod::class];
+    }
+
+    /**
+     * @param ClassMethod|Function_ $node
+     */
+    public function refactor(Node $node): ?Node
+    {
+        if ($this->isAtLeastPhpVersion($this->getPhpVersionFeature())) {
+            return null;
+        }
+
+        if ($this->shouldSkip($node)) {
+            return null;
+        }
+
+        if ($this->addDocBlock) {
+            /** @var PhpDocInfo|null $phpDocInfo */
+            $phpDocInfo = $node->getAttribute(AttributeKey::PHP_DOC_INFO);
+            if ($phpDocInfo === null) {
+                $phpDocInfo = $this->phpDocInfoFactory->createEmpty($node);
+            }
+
+            if ($node->returnType !== null) {
+                $type = $this->staticTypeMapper->mapPhpParserNodePHPStanType($node->returnType);
+                $phpDocInfo->changeReturnType($type);
+            }
+        }
+
+        $node->returnType = null;
+
+        return $node;
+    }
+
+    /**
+     * @param ClassMethod|Function_ $functionLike
+     */
+    private function shouldSkip(FunctionLike $functionLike): bool
+    {
+        if ($functionLike->returnType === null) {
+            return true;
+        }
+
+        return ! $this->shouldRemoveReturnDeclaration($functionLike);
+    }
+}
