+++ rules/downgrade/src/Rector/Property/DowngradeTypedPropertyRector.php
@@ -52,8 +52,8 @@ PHP
         return PhpVersionFeature::TYPED_PROPERTIES;
     }
 
-    public function shouldSkip(Property $property): bool
+    public function shouldRemoveProperty(Property $property): bool
     {
-        return false;
+        return true;
     }
 }
